clean up code duplication
