+++ zhaquirks/philips/rwl021.py
@@ -1,9 +1,7 @@
 """Phillips RWL021 device."""
-import logging
 
 from zigpy.profiles import zha, zll
-from zigpy.quirks import CustomCluster, CustomDevice
-import zigpy.types as t
+from zigpy.quirks import CustomDevice
 from zigpy.zcl.clusters.general import (
     Basic,
     BinaryInput,
@@ -16,10 +14,9 @@ from zigpy.zcl.clusters.general import (
     Scenes,
 )
 
+from . import PhillipsBasicCluster, PhillipsRemoteCluster
 from ..const import (
     ARGS,
-    COMMAND_ID,
-    BUTTON,
     CLUSTER_ID,
     COMMAND,
     COMMAND_OFF_WITH_EFFECT,
@@ -33,70 +30,13 @@ from ..const import (
     INPUT_CLUSTERS,
     LONG_PRESS,
     OUTPUT_CLUSTERS,
-    PRESS_TYPE,
     PROFILE_ID,
     SHORT_PRESS,
     TURN_OFF,
     TURN_ON,
-    ZHA_SEND_EVENT,
 )
 
 DIAGNOSTICS_CLUSTER_ID = 0x0B05  # decimal = 2821
-_LOGGER = logging.getLogger(__name__)
-
-
-class BasicCluster(CustomCluster, Basic):
-    """Phillips Basic cluster."""
-
-    attributes = Basic.attributes.copy()
-    attributes.update({0x0031: ("phillips", t.bitmap16)})
-
-    attr_config = {0x0031: 0x000B}
-
-    async def bind(self):
-        """Bind cluster."""
-        result = await super().bind()
-        await self.write_attributes(self.attr_config, manufacturer=0x100B)
-        return result
-
-
-class PhillipsCluster(CustomCluster):
-    """Phillips Basic cluster."""
-
-    cluster_id = 64512
-    name = "PhillipsCluster"
-    ep_attribute = "phillips_cluster"
-    attributes = {}
-    server_commands = {}
-    client_commands = {
-        0x0000: (
-            "notification",
-            (t.uint8_t, t.uint24_t, t.uint8_t, t.uint8_t, t.uint8_t, t.uint8_t),
-            False,
-        )
-    }
-    BUTTONS = {1: "on", 2: "up", 3: "down", 4: "off"}
-    PRESS_TYPES = {0: "press", 1: "hold", 2: "short_release", 3: "long_release"}
-
-    def handle_cluster_request(self, tsn, command_id, args):
-        """Handle the cluster command."""
-        _LOGGER.info(
-            "handle_cluster_request tsn: [%s] command id: %s - args: [%s]",
-            tsn,
-            command_id,
-            args,
-        )
-        button = self.BUTTONS.get(args[0], args[0])
-        press_type = self.PRESS_TYPES.get(args[2], args[2])
-
-        event_args = {
-            BUTTON: button,
-            PRESS_TYPE: press_type,
-            COMMAND_ID: command_id,
-            ARGS: args,
-        }
-        action = "{}_{}".format(button, press_type)
-        self.listener_event(ZHA_SEND_EVENT, action, event_args)
 
 
 class PhilipsRWL021(CustomDevice):
@@ -155,11 +95,11 @@ class PhilipsRWL021(CustomDevice):
             },
             2: {
                 INPUT_CLUSTERS: [
-                    BasicCluster,
+                    PhillipsBasicCluster,
                     PowerConfiguration.cluster_id,
                     Identify.cluster_id,
                     BinaryInput.cluster_id,
-                    PhillipsCluster,
+                    PhillipsRemoteCluster,
                 ],
                 OUTPUT_CLUSTERS: [Ota.cluster_id],
             },
